requirements.txt
blinker==1.9.0
boto3==1.37.13
botocore==1.37.13
click==8.1.8
colorama==0.4.6
docutils==0.16
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
pyasn1==0.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rsa==4.7.2
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
